index.md: make links out of docs-repo absolute
These pages don't exits yet, so let's link them to the fusetools.com
pages. Some might be moved to fuse-open.github.io eventually (or even
places like https://github.com/fuse-open/Fuse/releases), but making
them non-broken is a step in the right direction, and it makes it
easier to locate by grepping for fusetools.com in this repo.