build: restructure source-layout
It's easier to just pre-bake a directory for the static content than
copying the exact same files around. So let's just make the jump.

It's also a bit odd to have a split between assets and media, as media
is usually considered to be assets. So let's instead reflect the real
difference; the files in the assets-folder are *static*.