layout: move search into sidebar
This is a much more natural place to put it, and it makes it easier to
spot the search-field also. It also allows us to remove some strange
padding hacks that cause some strange rendering on mobile.

While we're at it, give it an id rather than a class, and set that right
on the form. This is simpler, and there should only be one anyway.